feat: add exceptions handling for EmployeeService

+++ src/main/java/com/clockin/service/EmployeeService.java
@@ -1,7 +1,10 @@
 package com.clockin.service;
 
+import com.clockin.exceptions.DataBaseException;
+import com.clockin.exceptions.EmployeeNotFoundException;
 import com.clockin.model.Employee;
 import com.clockin.repository.EmployeeRepository;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,27 +13,37 @@ import java.util.Optional;
 @Service
 public class EmployeeService {
 
-    private EmployeeRepository repository;
+    private final EmployeeRepository repository;
 
     public EmployeeService(EmployeeRepository repository) {
         this.repository = repository;
     }
 
     public List<Employee> getEmployees() {
-        List<Employee> employees = repository.findAll();
-        return employees;
+        return repository.findAll();
     }
 
     public Employee getEmployeeById(Long id) {
-        Optional<Employee> ep = repository.findById(id);
-        return ep.get();
+        Optional<Employee> employee = repository.findById(id);
+        if (employee.isPresent()) {
+            return employee.get();
+        } else
+            throw new EmployeeNotFoundException();
     }
 
-    public void registerEmployee(Employee employee){
+    public void registerEmployee(Employee employee) {
         repository.save(employee);
     }
 
-    public void deleteEmployeeById(Long employeeId){
-        repository.deleteById(employeeId);
+    public void deleteEmployeeById(Long employeeId) {
+        try {
+            if(repository.existsById(employeeId))
+            repository.deleteById(employeeId);
+            else {
+                throw new EmployeeNotFoundException();
+            }
+        } catch (DataIntegrityViolationException e) {
+            throw new DataBaseException("The employee cannot be deleted because they have associated workday records. Employee ID: " + employeeId);
+        }
     }
 }
